Uploadable profile pictures
Added a new field to users: profilePic, which has the name of an image used as their profile picture. Added new buttons to upload an image to be used as your profile picture. These buttons are only visible on the page of the account that is logged in.

Fixed an issue I was having where I was getting logged out every 5 seconds.